required modification
Made the script work properly -its done.
Made some changes in css - it is kind of done. If someone feel, can make it perfect.
With the script there are additional functions which can be worked on - like using the arrow keys to switch between the pics- but it is fine.